+++ tests/cbor_decode/test3_simple/src/main.c
@@ -124,7 +124,7 @@ uint8_t serial_rec_input2[] = {
 void test_pet(void)
 {
 	struct Pet pet;
-	size_t decode_len;
+	uint32_t decode_len;
 	uint8_t input[] = {
 		0x83, 0x82, 0x63, 0x66, 0x6f, 0x6f, 0x63, 0x62, 0x61, 0x72,
 		0x48, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08,
@@ -145,7 +145,7 @@ void test_pet(void)
 void test_serial1(void)
 {
 	struct Upload upload;
-	size_t decode_len;
+	uint32_t decode_len;
 	bool ret = cbor_decode_Upload(serial_rec_input1,
 			sizeof(serial_rec_input1), &upload, &decode_len);
 	zassert_true(ret, "decoding failed.");
@@ -168,7 +168,7 @@ void test_serial1(void)
 void test_serial2(void)
 {
 	struct Upload upload;
-	size_t decode_len;
+	uint32_t decode_len;
 	bool ret = cbor_decode_Upload(serial_rec_input2,
 			sizeof(serial_rec_input2), &upload, &decode_len);
 	zassert_true(ret, "decoding failed.");
